requirements.txt
Flask 
Jinja2 
Flask-SQLAlchemy 
psycopg2